refactor app component out of index.js
move app component to its own file in component folder
refactor trimBody function in helper.js
