+++ lib/helpers.js
@@ -1,13 +1,14 @@
 let helpers = {
 
-  trimBody : function(bodyText, maxLength) {
-    if (bodyText.length > maxLength) {
-      let trimmedString = bodyText.substr(0, maxLength);
-      trimmedString = trimmedString.substr(0, Math.min(maxLength, trimmedString.lastIndexOf(" ")));
-      return trimmedString + "...";
-    } else {
-      return bodyText;
-    }
+  trimBody(bodyText, maxLength) {
+    let bodyLength = bodyText.lastIndexOf(' ', maxLength);
+    let truncatedBody = bodyText.length > maxLength ? bodyText.substring(0, bodyLength) + '...' : bodyText
+    return truncatedBody
+  },
+
+
+  notFound({}) {
+    return <h1>Not Found!</h1>
   },
 
 }
